use server side fetching for remaining pages in settings/developer
