+++ apps/web/app/future/settings/(settings)/developer/webhooks/page.tsx
@@ -1,7 +1,14 @@
-import { _generateMetadata } from "app/_utils";
+import { _generateMetadata, getFixedT } from "app/_utils";
+import { getServerSession } from "next-auth";
+import { notFound } from "next/navigation";
 
-import { WebhooksViewAppDir } from "@calcom/features/webhooks/pages/webhooks-view";
+import { AUTH_OPTIONS } from "@calcom/feature-auth/lib/next-auth-options";
+import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
+import { CreateNewWebhookButton } from "@calcom/features/webhooks/components/CreateNewWebhookButton";
+import WebhooksView from "@calcom/features/webhooks/pages/webhooks-view";
 import { APP_NAME } from "@calcom/lib/constants";
+import { WebhookRepository } from "@calcom/lib/server/repository/webhook";
+import { UserPermissionRole } from "@calcom/prisma/enums";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -9,8 +16,31 @@ export const generateMetadata = async () =>
     (t) => t("add_webhook_description", { appName: APP_NAME })
   );
 
-const Page = () => {
-  return <WebhooksViewAppDir />;
+const Page = async () => {
+  const session = await getServerSession(AUTH_OPTIONS);
+  const user = session?.user;
+  if (!user) {
+    notFound();
+  }
+  const isAdmin = user.role === UserPermissionRole.ADMIN;
+
+  const t = await getFixedT(session?.user.locale || "en");
+
+  const data = await WebhookRepository.getAllWebhooksByUserId({
+    userId: user.id,
+    organizationId: user.profile?.organizationId,
+    userRole: user.role as UserPermissionRole,
+  });
+
+  return (
+    <SettingsHeader
+      title={t("webhooks")}
+      description={t("add_webhook_description", { appName: APP_NAME })}
+      CTA={data && data.webhookGroups.length > 0 ? <CreateNewWebhookButton isAdmin={isAdmin} /> : <></>}
+      borderInShellHeader={(data && data.profiles.length === 1) || !data?.webhookGroups?.length}>
+      <WebhooksView ssrProps={{ webhooks: data }} />
+    </SettingsHeader>
+  );
 };
 
 export default Page;
